requirements.txt
numpy
pandas
scipy
matplotlib
seaborn
scikit-learn
dvc
cookiecutter-data-science
pytest
fastapi
mlflow
evidently
sdv
deepchecks
joblib
pyarrow
pyyaml
xgboost
